+++ packages/fighting-components/button/src/button.vue
@@ -41,14 +41,16 @@
 </template>
 
 <script lang="ts" setup name="FButton">
-  import { computed, ref } from 'vue'
+  import { computed, ref, onMounted } from 'vue'
   import { Props, Emits } from './button'
+  // import { Ripples, ChangeColor } from '@fighting-design/fighting-utils'
   import { Ripples, ChangeColor } from '@fighting-design/fighting-utils'
   import type { ComputedRef, Ref } from 'vue'
   import type {
     buttonStyleInterface,
-    onClickInterface
-  } from '@fighting-design/fighting-type'
+    onClickInterface,
+    ordinaryFunctionInterface
+  } from './interface'
 
   const prop = defineProps(Props)
   const emit = defineEmits(Emits)
@@ -69,7 +71,8 @@
         blob,
         size,
         text,
-        circle
+        circle,
+        color
       } = prop
 
       return [
@@ -78,12 +81,12 @@
           [`f-button-${type}`]: type,
           [`f-button-${size}`]: size,
           'f-button-disabled': disabled || loading,
-          'f-button-simple': simple,
+          'f-button-simple': simple && !color,
           'f-button-circle': circle,
           'f-button-round': round,
           'f-button-block': block,
           'f-button-blob': blob,
-          'f-button-text': text
+          'f-button-text': text && !color
         }
       ]
     }
@@ -92,14 +95,6 @@
   const buttonStyle: ComputedRef<buttonStyleInterface | Object> = computed(
     (): buttonStyleInterface | Object => {
       const { fontSize, fontColor, color } = prop
-
-      if (color) {
-        const changeColor = new ChangeColor(color)
-        const light = changeColor.getLightColor(0.5)
-        const dark = changeColor.getDarkColor(0.5)
-
-        console.log(light, dark)
-      }
 
       return {
         fontSize,
@@ -110,7 +105,7 @@
   )
 
   const onClick: onClickInterface = (evt: PointerEvent): void => {
-    const { disabled, loading, ripples } = prop
+    const { disabled, loading, ripples, ripplesColor } = prop
 
     if (disabled || loading) {
       evt.preventDefault()
@@ -122,7 +117,7 @@
         evt,
         FButton.value as HTMLButtonElement,
         600,
-        prop.ripplesColor
+        ripplesColor
       )
       ripples.clickRipples()
     }
@@ -130,6 +125,7 @@
     emit('click', evt)
   }
 
+  // 左侧 icon
   const leftIconClass: ComputedRef<string> = computed<string>((): string => {
     const { loading, loadingIcon, leftIcon } = prop
 
@@ -137,5 +133,25 @@
       return `${loadingIcon || 'f-icon-loading'} f-loading-animation`
     }
     return leftIcon
+  })
+
+  // 自定义颜色
+  const customColor: ordinaryFunctionInterface = (): void => {
+    const { color } = prop
+    const changeColor: ChangeColor = new ChangeColor(color)
+    const light: string = changeColor.getLightColor(0.4)
+    const dark: string = changeColor.getDarkColor(0.1)
+    const node: HTMLButtonElement = FButton.value as HTMLButtonElement
+
+    node.addEventListener('mouseover', () => (node.style.background = light))
+    node.addEventListener('mousedown', () => (node.style.background = dark))
+    node.addEventListener('mouseup', () => (node.style.background = light))
+    node.addEventListener('mouseout', () => (node.style.background = color))
+  }
+
+  onMounted(() => {
+    if (prop.color) {
+      customColor()
+    }
   })
 </script>