+++ packages/fighting-components/icon/src/interface.d.ts
@@ -1,7 +1,3 @@
 export interface onClickInterface {
   (evt: PointerEvent): void
 }
-
-export interface ordinaryFunctionInterface {
-  (): void
-}