+++ packages/fighting-components/link/src/link.vue
@@ -30,7 +30,7 @@
 
 <script lang="ts" setup name="FLink">
   import { Props, Emits } from './link'
-  import type { onClickInterface } from '@fighting-design/fighting-type'
+  import type { onClickInterface } from './interface'
 
   const prop = defineProps(Props)
   const emit = defineEmits(Emits)