+++ packages/fighting-components/tag/src/tag.ts
@@ -1,4 +1,4 @@
-import type { tagSize, tagType } from '@fighting-design/fighting-type'
+import type { tagSize, tagType } from './interface'
 import type { PropType } from 'vue'
 
 export const Props = {