+++ packages/fighting-utils/changeColor.ts
@@ -1,7 +1,7 @@
-import type { changeColorInterface } from '@fighting-design/fighting-type'
+import type { changeColorInterface } from './type'
 
 /**
- * 根据传入的 hex
+ * 根据传入的 hex 计算出加深和减淡的颜色
  */
 export class ChangeColor implements changeColorInterface {
   color: string
@@ -43,7 +43,9 @@ export class ChangeColor implements changeColorInterface {
   getLightColor(level: number): string {
     const rgb: Array<string> = this.hexToRgb() as Array<string>
     for (let i: number = 0; i < rgb.length; i++) {
-      rgb[i] = Math.floor((255 - Number(rgb[i])) * level + Number(rgb[i])).toString(16)
+      rgb[i] = Math.floor(
+        (255 - Number(rgb[i])) * level + Number(rgb[i])
+      ).toString(16)
     }
     return this.rgbToHex(...rgb)
   }