+++ vite.config.ts
@@ -19,10 +19,12 @@ export default defineConfig({
   ],
   build: {
     lib: {
-      entry: resolve(__dirname, 'packages/fighting-components/index.ts'),
+      entry: resolve(__dirname, 'packages/fighting-design'),
       name: 'fightingDesign',
       formats: ['es'],
-      fileName: (format) => `index.${format}.js`
+      fileName: (format) => {
+        return 'index.js'
+      }
     },
     rollupOptions: {
       external: ['vue'],